server-app/requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.41.0
psycopg[binary]==3.3.3
python-dotenv==1.2.1
pytest==9.0.2
httpx==0.28.1
ruff==0.15.4
